Broke apart Write<T>
Write<T> was a generic that, via slow reflection, peeled type information off the type parameter defeating the purpose of the generic in the first place. Being that Write<T> has many references, is a critical section of networking code that produces packet bit arrays and was using said hacky stuff I've re-implemented it in a fashion that should be a design-wise and performance gain overall for the Packet class and the networking portion of the project as a whole.

(Installing VS14 atm so can't tell it this breaks anything. Also, as a side note interested in hearing from you guys about this emu project. A serverside in .Net/Mono could benefit the WoWemu community as a whole imo. Feel free to contact me.)